fix(daffio): the routing around api and guide docs is not correct
