+++ apps/daffio/src/app/core/sidebar/containers/sidebar-viewport/sidebar-viewport.component.html
@@ -2,7 +2,7 @@
   <daff-sidebar class="daffio-sidebar" (escapePressed)="close()">
     <button type="button" daff-button class="daffio-sidebar__button">Get Started</button>
 		<daffio-guide-nav [guideList]="sidebarContents$ | async"></daffio-guide-nav>
-    <a routerLink="/docs/api">API Index</a>
+    <a routerLink="/api">API Index</a>
   </daff-sidebar>
   <ng-content></ng-content>
 </daff-sidebar-viewport>
