+++ apps/daffio/src/app/docs/api-docs/api-routing.module.ts
@@ -16,6 +16,9 @@ export const apiRoutes: Routes = [
     resolve: {
       reference: DaffioApiListResolver,
     },
+    data: {
+      sidebarMode: 'side',
+    },
   },
   {
     path: '**',
